security: update .gitignore to exclude all sensitive information
- Add comprehensive .env file patterns (all variants)
- Exclude log files, backup files, and certificate/key files
- Add patterns for sensitive data files and temporary files
- Exclude BMAD and AI tool configs that may contain API keys
- Ensure no sensitive information can be accidentally committed

+++ app/Models/Account.php
@@ -250,7 +250,7 @@ class Account extends Model
             return null;
         }
 
-        return now()->diffInDays($this->payment_due_date, false);
+        return (int) now()->diffInDays($this->payment_due_date, false);
     }
 
     /**
